[Refactor] #29 - 리뷰 반영
