+++ Livith-iOS/Tuist/ProjectDescriptionHelpers/Module.swift
@@ -7,9 +7,9 @@
 
 import ProjectDescription
 
-// MARK: - ProjectName
+// MARK: - ProjectID
 
-public enum ProjectName: String, CaseIterable {
+public enum ProjectID: String, CaseIterable {
     case app = "App"
     case core = "Core"
     case shared = "Shared"
@@ -22,9 +22,9 @@ public enum ProjectName: String, CaseIterable {
     }
 }
 
-// MARK: - TargetName
+// MARK: - TargetID
 
-public enum TargetName {
+public enum TargetID {
     case app
     case core(CoreModule)
     case login(LoginModule)
@@ -74,33 +74,4 @@ public enum TargetName {
             return nil
         }
     }
-}
-
-// MARK: - Core Module
-
-public enum CoreModule: String {
-    case diContainer = "DIContainer"
-    case performanceMonitor = "PerformanceMonitor"
-    case routing = "Routing"
-    case persistence = "Persistence"
-    case network = "Network"
-}
-
-// MARK: - Shared Module
-
-public enum SharedModule: String {
-    case designSystem = "DesignSystem"
-}
-
-// MARK: - Login Module
-
-public enum LoginModule: String {
-    case loginFeature = "LoginFeature"
-}
-
-// MARK: - External Dependency
-
-public enum ExternalDependency: String {
-    case alamofire = "Alamofire"
-    case kingfisher = "Kingfisher"
 }
